backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
pydantic==2.9.2
alembic==1.13.2